+++ src/autoAnimations.js
@@ -21,7 +21,6 @@ import shatterAuto from "./animation-functions/shatter.js";
 import onTargetSpells from "./animation-functions/healing-spells.js";
 import magicMissile from "./animation-functions/magic-missile.js";
 import arrowOptionExplode from "./animation-functions/arrow.js";
-import explodeTemplate from "./animation-functions/explosion-template.js";
 import explodeOnTarget from "./animation-functions/explode-ontarget.js";
 import castOnSelf from "./animation-functions/shield.js";
 import selfCast from "./animation-functions/emanations.js";
@@ -30,8 +29,8 @@ import huntersMark from "./animation-functions/hunters-mark.js";
 import bardicInspiration from "./animation-functions/bardic-inspiration.js";
 import mistyStep from "./animation-functions/misty-step.js";
 import unarmedStrike from "./animation-functions/unarmed-strike.js";
-import breathWeapon from "./animation-functions/breath-weapons.js";
 import mistyStepOld from "./animation-functions/misty-step-old.js";
+import templateEffects from "./animation-functions/template-effects.js";
 
 import { AALayer } from "./canvas-animation/AutoAnimationsLayer.js";
 import ImagePicker from "./ImagePicker.js";
@@ -42,23 +41,25 @@ const log = () => { };
 
 function registerLayer() {
     if (game.data.version === "0.7.9" || game.data.version === "0.7.10") {
-        const layers = mergeObject(Canvas.layers, {
+        CONFIG.Canvas.layers = mergeObject(CONFIG.Canvas.layers, {
             autoanimations: AALayer
-        });
-        Object.defineProperty(Canvas, 'layers', {
-            get: function () {
-                return layers
-            }
         });
     } else {
-        const layers = foundry.utils.mergeObject(Canvas.layers, {
+        CONFIG.Canvas.layers = foundry.utils.mergeObject(CONFIG.Canvas.layers, {
             autoanimations: AALayer
         });
+    }
+
+    // workaround for other modules
+    if (!Object.is(Canvas.layers, CONFIG.Canvas.layers)) {
+        console.error('Possible incomplete layer injection by other module detected! Trying workaround...')
+
+        const layers = Canvas.layers
         Object.defineProperty(Canvas, 'layers', {
             get: function () {
-                return layers
+                return foundry.utils.mergeObject(layers, CONFIG.Canvas.layers)
             }
-        });
+        })
     }
 }
 
@@ -460,12 +461,9 @@ async function specialCaseAnimations(msg) {
                 mistyStep(handler);
             }
             break;
-        case handler.animType === "t14":
-            Hooks.once("createMeasuredTemplate", () => {
-                if (handler.itemSound) {
-                    itemSound(handler);
-                }
-                breathWeapon(handler);
+        case (handler.animType === "t8" && handler.animOverride):
+            Hooks.once("createMeasuredTemplate", (msg) => {
+                templateEffects(handler, msg);
             })
             break;
     }
@@ -498,11 +496,6 @@ function revItUp5eCore(msg) {
                         mistyStep(handler);
                     }
                     break;
-                case handler.animType === "t14":
-                    Hooks.once("createMeasuredTemplate", () => {
-                        breathWeapon(handler);
-                    })
-                    break;
             }
         }
         return;
@@ -515,6 +508,13 @@ function revItUp5eCore(msg) {
     }
     if (handler.itemNameIncludes("bardic inspiration") || handler.itemNameIncludes(game.i18n.format("AUTOANIM.bardicInspiration").toLowerCase())) {
         bardicInspiration(handler);
+    }
+
+    if (handler.animType === "t8" && handler.animOverride) {
+        Hooks.once("createMeasuredTemplate", (msg) => {
+            templateEffects(handler, msg);
+        });
+        return;
     }
 
     if (game.user.id === msg.user.id) {
@@ -557,22 +557,6 @@ function revItUp5eCore(msg) {
                                         shatterAuto(handler);
                                     })
                                     break;
-                                case (handler.itemNameIncludes("grenade")):
-                                case (handler.itemNameIncludes("bomb")):
-                                case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemGrenade").toLowerCase()):
-                                case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemBomb").toLowerCase()):
-                                case (handler.animType === "t8"):
-                                    Hooks.once("createMeasuredTemplate", () => {
-                                        explodeTemplate(handler);
-                                    })
-                                    break;
-                                case handler.animType === "t14":
-                                    Hooks.once("createMeasuredTemplate", () => {
-                                        breathWeapon(handler);
-                                    })
-                                    break;
-                                default:
-                                    revItUp(handler);
                             }
                             break;
                         case (false):
@@ -581,10 +565,7 @@ function revItUp5eCore(msg) {
                     }
                 }
 
-            } else {
-                log("MRE is NOT active");
-                revItUp(handler);
-            }
+            } else { log("MRE is NOT active"); revItUp(handler); }
         }
         switch (true) {
             case (handler.itemTypeIncludes("spell") && handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemShield").toLowerCase())):
@@ -596,7 +577,7 @@ function revItUp5eCore(msg) {
         if (rollType.includes("damage")) {
             log("damage roll");
         } else
-            if (rollType.includes("attack")) {
+            if (rollType.includes("attack") || !handler.hasAttack) {
                 if (handler.itemSound) {
                     itemSound(handler);
                 }
@@ -614,11 +595,7 @@ function revItUp5eCore(msg) {
                         itemSound(handler);
                     }
                     switch (true) {
-                        case handler.animType === "t14":
-                            Hooks.once("createMeasuredTemplate", () => {
-                                breathWeapon(handler);
-                            })
-                            break;
+                        /*
                         case (handler.itemNameIncludes("cure", "wound")):
                         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemCureWounds").toLowerCase()):
                         case (handler.itemNameIncludes("heal", "word")):
@@ -649,6 +626,7 @@ function revItUp5eCore(msg) {
                         //case (handler.itemNameIncludes("siege")):
                         //rangedWeapons(handler)
                         //break;
+                        */
                         case (handler.itemNameIncludes("thunderwave")):
                         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemThunderwave").toLowerCase()):
                             Hooks.once("createMeasuredTemplate", () => {
@@ -659,15 +637,6 @@ function revItUp5eCore(msg) {
                         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemShatter").toLowerCase()):
                             Hooks.once("createMeasuredTemplate", () => {
                                 shatterAuto(handler);
-                            })
-                            break;
-                        case (handler.itemNameIncludes("grenade")):
-                        case (handler.itemNameIncludes("bomb")):
-                        case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemGrenade").toLowerCase()):
-                        case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemBomb").toLowerCase()):
-                        case (handler.animType === "t8"):
-                            Hooks.once("createMeasuredTemplate", () => {
-                                explodeTemplate(handler);
                             })
                             break;
                     }
@@ -688,6 +657,30 @@ async function revItUp(handler) {
     if (handler.itemSound) {
         itemSound(handler);
     }
+
+    if (game.modules.get("midi-qol")?.active) { } else {
+        switch (game.system.id) {
+            case "dnd5e":
+                if (game.modules.get("mars-5e")?.active) {
+                    if (handler.animType === "t8" && handler.animOverride) {
+                        templateEffects(handler);
+                        return;
+                    }
+                } else if (handler.animType === "t8" && handler.animOverride) {
+                    Hooks.once("createMeasuredTemplate", (msg) => {
+                        templateEffects(handler, msg);
+                    });
+                    return;
+                }
+                break;
+            default:
+                if (handler.animType === "t8" && handler.animOverride) {
+                    templateEffects(handler);
+                    return;
+                }
+        }
+    }
+
     switch (true) {
         // Use xxx in Item Source Field to exclude an item for On-Use customization
         case (handler.itemIncludes("xxx")):
@@ -695,9 +688,6 @@ async function revItUp(handler) {
             break;
         case ((handler.animType === "t9") && handler.animOverride):
             explodeOnTarget(handler);
-            break;
-        case ((handler.animType === "t8") && handler.animOverride):
-            explodeTemplate(handler);
             break;
         case ((handler.animType === "t10") && handler.animOverride):
             selfCast(handler);
@@ -776,7 +766,17 @@ async function revItUp(handler) {
             break;
         case (handler.itemNameIncludes("thunderwave")):
         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemThunderwave").toLowerCase()):
-            thunderwaveAuto(handler);
+            switch (true) {
+                case (game.modules.get("midi-qol")?.active && (handler.autoDamage === "none")):
+                    console.log(handler.autoDamage);
+                    thunderwaveAuto(handler);
+                    break;
+                default:
+                    Hooks.once("createMeasuredTemplate", () => {
+                        console.log(handler.autoDamage);
+                        thunderwaveAuto(handler);
+                    })
+            }
             break;
         case (handler.itemNameIncludes("shatter")):
         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemShatter").toLowerCase()):
@@ -814,12 +814,6 @@ async function revItUp(handler) {
         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemShield").toLowerCase()):
             castOnSelf(handler);
             break;
-        case (handler.itemNameIncludes("grenade")):
-        case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemGrenade").toLowerCase()):
-        case (handler.itemNameIncludes("bomb")):
-        case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemBomb").toLowerCase()):
-            explodeTemplate(handler);
-            break;
         case (handler.itemNameIncludes("bite") && !handler.itemNameIncludes("frost")):
         case (handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemBite").toLowerCase()) && !handler.itemNameIncludes("frost")):
         case (handler.itemNameIncludes("claw")):
@@ -830,25 +824,11 @@ async function revItUp(handler) {
         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemHM").toLowerCase()):
             huntersMark(handler)
             break;
-        /*
-        case (handler.itemNameIncludes("potion", "heal")):
-            castOnSelf(handler);
-            break;
-        case (handler.itemNameIncludes("second", "wind")):
-            castOnSelf(handler);
-            break;
-        */
         case handler.itemNameIncludes("unarmed strike"):
         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemUnarmedStrike").toLowerCase()):
         case handler.itemNameIncludes("flurry of blows"):
         case handler.itemNameIncludes(game.i18n.format("AUTOANIM.itemFlurryBlows").toLowerCase()):
             unarmedStrike(handler);
-            break;
-        case handler.animType === "t14":
-            if (game.modules.get("midi-qol")?.active) {
-            } else {
-                breathWeapon(handler);
-            }
             break;
     }
 }