+++ src/system-handlers/sw5e-handler.js
@@ -46,7 +46,9 @@ export default class SW5eHandler {
         this._explodeSound = this._flags.allSounds?.explosion?.audioExplodeEnabled ?? false;
         this._spellLoops = this._flags?.spellOptions?.spellLoops ?? 1;
         this._divineSmite = this._flags.divineSmite ?? "";
-      
+        this._templates = this._flags.templates ?? "";
+        this._item = this._actorToken.actor.items?.get(itemId) ?? "";
+
         //console.log(this._animName);
         this._animNameFinal;
         switch (true) {
@@ -233,6 +235,14 @@ export default class SW5eHandler {
         return this._divineSmite;
     }
     
+    get templates() {
+        return this._templates;
+    }
+    
+    get hasAttack() {
+        return this._item.hasAttack;
+    }
+
     getDistanceTo(target) {
         var x, x1, y, y1, d, r, segments = [], rdistance, distance;
         for (x = 0; x < this._actorToken.data.width; x++) {